% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{generated}
\alias{generated}
\title{Docs for the \code{generated} data set}
\format{An object of class \code{data.frame} with 10 rows and 3 columns.}
\usage{
generated
}
\description{
Docs for the \code{generated} data set
}
\keyword{datasets}
