% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{Docs for the \code{mydata} data set}
\format{An object of class \code{data.frame} with 10 rows and 3 columns.}
\usage{
mydata
}
\description{
Docs for the \code{mydata} data set
}
\keyword{datasets}
