% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parrotmetric.R
\name{parametric_equation}
\alias{parametric_equation}
\title{This is a title}
\usage{
parametric_equation(x, y, t0, t1)
}
\arguments{
\item{x}{Functions representing the x values in terms of t.}

\item{y}{Functions representing the y values in terms of t.}

\item{t0}{The initial value of t}

\item{t1}{The final value of t}
}
\value{
An object of \code{parametric_equation}
}
\description{
Construct a parametric equation object
}
\details{
Description goes here.
}
\section{Beware!}{


Don't use this function for mission critical things.
}

\examples{
# construct a parametric equation object
\dontrun{
obj <- parametric_equation(sin, cos, 0, 2 * pi)
obj
}

# See .... for ways of using it
}
\seealso{
See \code{\link{median}} for something totally different
}
